libpcap: Add patch to fix instability of SONAME value in binary.
libpcap in upstream has build target "libpcap.so" in Makefile.in to generate libpcap binary and it sets SONAME to "libpcap.so.1".

But by historical reason[1], debian specifies its libpcap's SONAME to "libpcap.so.0.8".
And to do that, debian replaces the "libpcap.so" target with "$(SHAREDLIB)" target, which generates binary and sets SONAME to "libpcap.so.0.8".

We found that the target "libpcap.so" is rarely executed and soname is set to "libpcap.so.1" (We got this result once in 20 times build).
This is because the target "install-shared-so" depends on the target "libpcap.so" and especially in parallel building, the target "libpcap.so" is executed according to the timing of generating object files.

To ensure the target "libpcap.so" is never executed, we deleted it and modified to make the "$(SHAREDLIB)" target always executed.

[1]:https://people.debian.org/~rfrancoise/libpcap-faq.html