python3: resolve ncursesw host contamination
When the libncursesw5-dev package is installed on the host, its header files
are unintentionally used to build python3. Poky/OE fixed this issue in the
following commit.

  commit 39970f583dddf04139a039cbe04e62b7236f0481
  python3: resolve ncurses host contamination

I found this issue when I tried to use ABI version 6 of ncurses in
poky/meta-debian. For ABI version 5 (default in Poky/OE and meta-debian),the
build looks successful in most cases although this undesirable contamination
actually occurs.