classes: kernel-cve-check: Disable KERNEL_CVE_CHECK_INCLUDE_IGNORE
When KERNEL_CVE_CHECK_INCLUDE_IGNORE is set to 1, it adds
--include-ignore option to report_affected.py. This means tell
reporte_affected.py to show CVEs which are marked as ignore. Therefore,
if ignored CVE is not fixed in NVD database, it will be reported as not
fixed.  This is not expected result because cip-kernel-sec decided to
ignore the CVE.
So, disable KERNEL_CVE_CHECK_INCLUDE_IGNORE flag to not show CVEs which
are marked as ignore.

If we don't have any problems with this setting, we can remove
KERNEL_CVE_CHECK_INCLUDE_IGNORE and KERNEL_CVE_CHECK_SHOW_IGNORE_INFO
related code in the future.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>