Add new image emlinux-image-minimal
This image support to create small footprint image.
It uses busybox for basic commands instead of coreutils, util-linux, and
so on.

Features:

- Use busybox commands instead of coreutils
- Use sysvinit as init system
- Remove packages from emlinux-image-base

Limitations:

- This image disrupts the Debian system features expected by ISAR
- When you want to update an image, you need to run "bitbake
emlinux-image-minimal -c cleanall" first
- SDK doesn't support minimal image

To build emlinux-image-minimal, it runs following steps.

1. Create emlinux-image-base
2. Replace commands to busybox from coreutils, util-linux, and etc
3. Remove packages to reduce footprint

Options:

There are some options to reduce footprint.

- EMLINUX_IMAGE_MINIMAL_REMOVE_PERL
If set to 1, remove perl packages.

- EMLINUX_IMAGE_MINIMAL_REMOVE_BOOT_DIR
If set to 1, remove files in /boot directory.
EMLinux's rootfs contains /boot in it, however if your system boot
from u-boot, /boot is not needed so this option remove files in
/boot to reduce footprint.

- EMLINUX_IMAGE_MINIMAL_REMOVE_MAN
If set to 1, remove man files.

- EMLINUX_IMAGE_MINIMAL_REMOVE_DTB
If set to 1, remove dtb files in /usr/lib/linux-image-<kernel version>.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>