recipes-core/cve-update: Check vulnStatus key before use
According to the NVD CVE API Schema[1], it seems as if the vulnStatus
always exists in a response. But we found CVE-2025-2682 does not
contain it as of 2025-03-28[1].
This key has been added as of 2025-04-01 however it would be nice to
check the value is in the API result before use to prevent future error.

1: https://github.com/miraclelinux/meta-emlinux/pull/452/commits/babb7fad8800f301dc8cac6e1953589d6bd7ec83

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>