conf/cve: Add CVEs to ignore in EMLinux
Following bugs can be ignored.

- CVE-2021-0399
This is not mainline kernel bug.

- CVE-2021-29256
This is not mainline kernel bug.

- CVE-2021-3492
This is not mainline kernel bug.

- CVE-2021-39802
This is not mainline kernel bug.

- CVE-2022-2327
This bug is only affeceted to linux 5.10.

- CVE-2022-36397
This is not mainline kernel bug.

- CVE-2023-0030
This bug is false positive.

- CVE-2023-26242
This bug is false positive.

- CVE-2023-3079
This is not mainline kernel bug.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>